fix: 채팅 목록 컨테이너 분리
- 서버에서 채팅 응답 시 페이지 전체가 리렌더링 되던 문제 개선
